Estimate asymptotic Hill diversity

Computes an estimate of the true community diversity based on a finite sample assuming that individuals are randomly and independently sampled, after Chao and Jost 2015. SpadeR source code modified to match our use of \(\ell\) instead of q to parameterize Hill diversity. Code taken from SpadeR:::Chao_Hill_abu()From the R package SpadeR 0.1.1 by Anne Chao, K. H. Ma, T. C. Hsieh and Chun-Huo Chiu

Chao_Hill_abu(ab, l, q = NULL)

Arguments

ab

A numeric vector of species abundances or relative abundances.

l

Scaling exponent for the mean, can be any real number.

q

Scalar, traditional Hill number scaling exponent, q = 1-l where l is the scaling parameter for the generalized mean. Can be any real number.

Value

Scalar estimate of Hill diversity given \(\ell\)

References

Chao A, Jost L (2015). “Estimating diversity and entropy profiles via discovery rates of new species.” Methods in Ecology and Evolution, 6(8), 873--882. ISSN 2041210X, doi: 10.1111/2041-21 .

Examples

# generate sample
abs <- sample_infinite(fit_SAD(rich = 50, simpson = 20)[[3]], 150)

# estimate true Hill diversity
Chao_Hill_abu(abs, l = 1) # Chao1 estimate of richness lower bound
#> [1] 55.90667
Chao_Hill_abu(abs, l = 0) # Asymptotic estimate of Hill-Shannon diversity
#> [1] 32.85384
Chao_Hill_abu(abs, l = -1) # Asymptotic estimate of Hill-Simpson diversity
#> [1] 19.40104