MeanRarity: A package for computing Hill diversity under the "mean rarity" framework.

The MeanRarity package provides a suite of functions related to computing, estimating, and visualizing Hill diversity: For references on mean rarity, see (Roswell et al. 2021) , (Patil and Taillie 1982) , and (Jost 2006)

Computing Hill diversity

rarity computes Hill diversity as parameterized by (Roswell et al. 2021) , by computing the generalized weighted mean of species rarities (i.e. the reciprocal of relative abundance). The "link" functions for the generalized mean are pfun and its inverse ipfun

Simulating Species Abundance Distributions

(Roswell et al. 2021) developed a method to simulate a species abundance distribution (SAD) based on the [Hill] diversity of the assemblage. This method fits a discretization (based on quantile values) of a continuous parametric distribution (currently gamma or log-normal) based on the known, simulated richness and (currently) Hill-Simpson diversity. The function to do this is called fit_SAD

Visualizing Mean Rarity

The code for rarity balance plots, also in the web app https://mean-rarity.shinyapps.io/rshiny_app1/ is also contained in this package. rarity_plot makes rarity balance plots. radplot makes plots of rank-abundance distributions, with or without log-transforming abundances

References

Jost L (2006). “Entropy and diversity.” Oikos, 113, 363--375. doi: 10.1111/j.2006.0030-1299.14714.x .

Patil GP, Taillie C (1982). “Diversity as a concept and its measurement.” Journal of the American Statistical Association, 77(379), 548--561. doi: 10.1080/01621459.1982.10477845 , http://www.tandfonline.com/loi/uasa20%5Cnhttp://dx.doi.org/10.1080/01621459.1982.10477845%5Cnhttp://www.tandfonline.com/.

Roswell M, Dushoff J, Winfree R (2021). “A conceptual guide to measuring species diversity.” Oikos, 130(3), 321--338. ISSN 0030-1299, doi: 10.1111/oik.07202 , https://onlinelibrary.wiley.com/doi/10.1111/oik.07202.

Author

Maintainer: Michael Roswell mike.roswell@gmail.com (ORCID)

Authors:

Jonathan Dushoff jdushoff@gmail.com (ORCID)