Generate Hill diversity profile

Compute observed Hill diversity profile based on an abundance vector over a range of scaling exponent values.

divpro(ab, ell_low = -1, ell_hi = 1, by = 0.001)

Arguments

ab
A numeric vector of species abundances or relative abundances.

ell_low
Scalar, minimum scaling exponent for diversity profile.

ell_hi
Scalar, maximum scaling exponent for diversity profile.

by
Scalar, size of step along scaling exponent continuum.

Value

Dataframe with the scaling exponent ell and corresponding Hill diversity d

Details

Hill diversity can be viewed as a continuous function of the scaling exponent \(\ell\) and the relative abundance distribution. As \(\ell\) increases, so does the emphasis on rare species. It is traditional to view the profile across \(\ell = [-1, 1]\) or \(\ell = [-2, 1]\), and other authors have visualized this with low values of \(\ell\) at the right instead of left.

Examples

divpro(c(20,8,5,4,2,1))

#>         ell        d
#> 1    -1.000 3.137255
#> 2    -0.999 3.137860
#> 3    -0.998 3.138466
#> 4    -0.997 3.139072
#> 5    -0.996 3.139679
#> 6    -0.995 3.140286
#> 7    -0.994 3.140893
#> 8    -0.993 3.141501
#> 9    -0.992 3.142110
#> 10   -0.991 3.142719
#> 11   -0.990 3.143328
#> 12   -0.989 3.143938
#> 13   -0.988 3.144548
#> 14   -0.987 3.145159
#> 15   -0.986 3.145771
#> 16   -0.985 3.146382
#> 17   -0.984 3.146995
#> 18   -0.983 3.147607
#> 19   -0.982 3.148221
#> 20   -0.981 3.148834
#> 21   -0.980 3.149448
#> 22   -0.979 3.150063
#> 23   -0.978 3.150678
#> 24   -0.977 3.151294
#> 25   -0.976 3.151910
#> 26   -0.975 3.152526
#> 27   -0.974 3.153143
#> 28   -0.973 3.153761
#> 29   -0.972 3.154379
#> 30   -0.971 3.154997
#> 31   -0.970 3.155616
#> 32   -0.969 3.156236
#> 33   -0.968 3.156856
#> 34   -0.967 3.157476
#> 35   -0.966 3.158097
#> 36   -0.965 3.158718
#> 37   -0.964 3.159340
#> 38   -0.963 3.159962
#> 39   -0.962 3.160585
#> 40   -0.961 3.161208
#> 41   -0.960 3.161832
#> 42   -0.959 3.162456
#> 43   -0.958 3.163081
#> 44   -0.957 3.163706
#> 45   -0.956 3.164332
#> 46   -0.955 3.164958
#> 47   -0.954 3.165585
#> 48   -0.953 3.166212
#> 49   -0.952 3.166840
#> 50   -0.951 3.167468
#> 51   -0.950 3.168097
#> 52   -0.949 3.168726
#> 53   -0.948 3.169355
#> 54   -0.947 3.169985
#> 55   -0.946 3.170616
#> 56   -0.945 3.171247
#> 57   -0.944 3.171879
#> 58   -0.943 3.172511
#> 59   -0.942 3.173143
#> 60   -0.941 3.173776
#> 61   -0.940 3.174410
#> 62   -0.939 3.175044
#> 63   -0.938 3.175678
#> 64   -0.937 3.176313
#> 65   -0.936 3.176949
#> 66   -0.935 3.177585
#> 67   -0.934 3.178221
#> 68   -0.933 3.178858
#> 69   -0.932 3.179496
#> 70   -0.931 3.180134
#> 71   -0.930 3.180772
#> 72   -0.929 3.181411
#> 73   -0.928 3.182051
#> 74   -0.927 3.182691
#> 75   -0.926 3.183331
#> 76   -0.925 3.183972
#> 77   -0.924 3.184614
#> 78   -0.923 3.185256
#> 79   -0.922 3.185898
#> 80   -0.921 3.186541
#> 81   -0.920 3.187184
#> 82   -0.919 3.187828
#> 83   -0.918 3.188473
#> 84   -0.917 3.189118
#> 85   -0.916 3.189763
#> 86   -0.915 3.190409
#> 87   -0.914 3.191056
#> 88   -0.913 3.191703
#> 89   -0.912 3.192350
#> 90   -0.911 3.192998
#> 91   -0.910 3.193647
#> 92   -0.909 3.194296
#> 93   -0.908 3.194945
#> 94   -0.907 3.195595
#> 95   -0.906 3.196246
#> 96   -0.905 3.196897
#> 97   -0.904 3.197548
#> 98   -0.903 3.198200
#> 99   -0.902 3.198853
#> 100  -0.901 3.199506
#> 101  -0.900 3.200160
#> 102  -0.899 3.200814
#> 103  -0.898 3.201468
#> 104  -0.897 3.202123
#> 105  -0.896 3.202779
#> 106  -0.895 3.203435
#> 107  -0.894 3.204092
#> 108  -0.893 3.204749
#> 109  -0.892 3.205407
#> 110  -0.891 3.206065
#> 111  -0.890 3.206723
#> 112  -0.889 3.207383
#> 113  -0.888 3.208042
#> 114  -0.887 3.208703
#> 115  -0.886 3.209363
#> 116  -0.885 3.210025
#> 117  -0.884 3.210686
#> 118  -0.883 3.211349
#> 119  -0.882 3.212012
#> 120  -0.881 3.212675
#> 121  -0.880 3.213339
#> 122  -0.879 3.214003
#> 123  -0.878 3.214668
#> 124  -0.877 3.215333
#> 125  -0.876 3.215999
#> 126  -0.875 3.216666
#> 127  -0.874 3.217333
#> 128  -0.873 3.218000
#> 129  -0.872 3.218668
#> 130  -0.871 3.219337
#> 131  -0.870 3.220006
#> 132  -0.869 3.220676
#> 133  -0.868 3.221346
#> 134  -0.867 3.222016
#> 135  -0.866 3.222687
#> 136  -0.865 3.223359
#> 137  -0.864 3.224031
#> 138  -0.863 3.224704
#> 139  -0.862 3.225377
#> 140  -0.861 3.226051
#> 141  -0.860 3.226726
#> 142  -0.859 3.227400
#> 143  -0.858 3.228076
#> 144  -0.857 3.228752
#> 145  -0.856 3.229428
#> 146  -0.855 3.230105
#> 147  -0.854 3.230783
#> 148  -0.853 3.231461
#> 149  -0.852 3.232139
#> 150  -0.851 3.232818
#> 151  -0.850 3.233498
#> 152  -0.849 3.234178
#> 153  -0.848 3.234859
#> 154  -0.847 3.235540
#> 155  -0.846 3.236222
#> 156  -0.845 3.236904
#> 157  -0.844 3.237587
#> 158  -0.843 3.238270
#> 159  -0.842 3.238954
#> 160  -0.841 3.239638
#> 161  -0.840 3.240323
#> 162  -0.839 3.241009
#> 163  -0.838 3.241695
#> 164  -0.837 3.242382
#> 165  -0.836 3.243069
#> 166  -0.835 3.243756
#> 167  -0.834 3.244445
#> 168  -0.833 3.245133
#> 169  -0.832 3.245823
#> 170  -0.831 3.246512
#> 171  -0.830 3.247203
#> 172  -0.829 3.247894
#> 173  -0.828 3.248585
#> 174  -0.827 3.249277
#> 175  -0.826 3.249970
#> 176  -0.825 3.250663
#> 177  -0.824 3.251356
#> 178  -0.823 3.252050
#> 179  -0.822 3.252745
#> 180  -0.821 3.253440
#> 181  -0.820 3.254136
#> 182  -0.819 3.254833
#> 183  -0.818 3.255529
#> 184  -0.817 3.256227
#> 185  -0.816 3.256925
#> 186  -0.815 3.257623
#> 187  -0.814 3.258322
#> 188  -0.813 3.259022
#> 189  -0.812 3.259722
#> 190  -0.811 3.260423
#> 191  -0.810 3.261124
#> 192  -0.809 3.261826
#> 193  -0.808 3.262529
#> 194  -0.807 3.263231
#> 195  -0.806 3.263935
#> 196  -0.805 3.264639
#> 197  -0.804 3.265344
#> 198  -0.803 3.266049
#> 199  -0.802 3.266754
#> 200  -0.801 3.267461
#> 201  -0.800 3.268168
#> 202  -0.799 3.268875
#> 203  -0.798 3.269583
#> 204  -0.797 3.270291
#> 205  -0.796 3.271000
#> 206  -0.795 3.271710
#> 207  -0.794 3.272420
#> 208  -0.793 3.273131
#> 209  -0.792 3.273842
#> 210  -0.791 3.274554
#> 211  -0.790 3.275266
#> 212  -0.789 3.275979
#> 213  -0.788 3.276693
#> 214  -0.787 3.277407
#> 215  -0.786 3.278121
#> 216  -0.785 3.278837
#> 217  -0.784 3.279552
#> 218  -0.783 3.280269
#> 219  -0.782 3.280986
#> 220  -0.781 3.281703
#> 221  -0.780 3.282421
#> 222  -0.779 3.283140
#> 223  -0.778 3.283859
#> 224  -0.777 3.284579
#> 225  -0.776 3.285299
#> 226  -0.775 3.286020
#> 227  -0.774 3.286741
#> 228  -0.773 3.287463
#> 229  -0.772 3.288186
#> 230  -0.771 3.288909
#> 231  -0.770 3.289633
#> 232  -0.769 3.290357
#> 233  -0.768 3.291082
#> 234  -0.767 3.291807
#> 235  -0.766 3.292533
#> 236  -0.765 3.293260
#> 237  -0.764 3.293987
#> 238  -0.763 3.294714
#> 239  -0.762 3.295443
#> 240  -0.761 3.296171
#> 241  -0.760 3.296901
#> 242  -0.759 3.297631
#> 243  -0.758 3.298361
#> 244  -0.757 3.299093
#> 245  -0.756 3.299824
#> 246  -0.755 3.300557
#> 247  -0.754 3.301289
#> 248  -0.753 3.302023
#> 249  -0.752 3.302757
#> 250  -0.751 3.303492
#> 251  -0.750 3.304227
#> 252  -0.749 3.304963
#> 253  -0.748 3.305699
#> 254  -0.747 3.306436
#> 255  -0.746 3.307173
#> 256  -0.745 3.307911
#> 257  -0.744 3.308650
#> 258  -0.743 3.309389
#> 259  -0.742 3.310129
#> 260  -0.741 3.310870
#> 261  -0.740 3.311611
#> 262  -0.739 3.312352
#> 263  -0.738 3.313094
#> 264  -0.737 3.313837
#> 265  -0.736 3.314581
#> 266  -0.735 3.315325
#> 267  -0.734 3.316069
#> 268  -0.733 3.316814
#> 269  -0.732 3.317560
#> 270  -0.731 3.318306
#> 271  -0.730 3.319053
#> 272  -0.729 3.319801
#> 273  -0.728 3.320549
#> 274  -0.727 3.321297
#> 275  -0.726 3.322047
#> 276  -0.725 3.322796
#> 277  -0.724 3.323547
#> 278  -0.723 3.324298
#> 279  -0.722 3.325049
#> 280  -0.721 3.325802
#> 281  -0.720 3.326554
#> 282  -0.719 3.327308
#> 283  -0.718 3.328062
#> 284  -0.717 3.328816
#> 285  -0.716 3.329572
#> 286  -0.715 3.330327
#> 287  -0.714 3.331084
#> 288  -0.713 3.331841
#> 289  -0.712 3.332598
#> 290  -0.711 3.333357
#> 291  -0.710 3.334115
#> 292  -0.709 3.334875
#> 293  -0.708 3.335635
#> 294  -0.707 3.336395
#> 295  -0.706 3.337156
#> 296  -0.705 3.337918
#> 297  -0.704 3.338681
#> 298  -0.703 3.339444
#> 299  -0.702 3.340207
#> 300  -0.701 3.340971
#> 301  -0.700 3.341736
#> 302  -0.699 3.342502
#> 303  -0.698 3.343268
#> 304  -0.697 3.344034
#> 305  -0.696 3.344801
#> 306  -0.695 3.345569
#> 307  -0.694 3.346338
#> 308  -0.693 3.347107
#> 309  -0.692 3.347876
#> 310  -0.691 3.348647
#> 311  -0.690 3.349417
#> 312  -0.689 3.350189
#> 313  -0.688 3.350961
#> 314  -0.687 3.351734
#> 315  -0.686 3.352507
#> 316  -0.685 3.353281
#> 317  -0.684 3.354056
#> 318  -0.683 3.354831
#> 319  -0.682 3.355606
#> 320  -0.681 3.356383
#> 321  -0.680 3.357160
#> 322  -0.679 3.357937
#> 323  -0.678 3.358716
#> 324  -0.677 3.359495
#> 325  -0.676 3.360274
#> 326  -0.675 3.361054
#> 327  -0.674 3.361835
#> 328  -0.673 3.362616
#> 329  -0.672 3.363398
#> 330  -0.671 3.364181
#> 331  -0.670 3.364964
#> 332  -0.669 3.365748
#> 333  -0.668 3.366532
#> 334  -0.667 3.367317
#> 335  -0.666 3.368103
#> 336  -0.665 3.368889
#> 337  -0.664 3.369676
#> 338  -0.663 3.370463
#> 339  -0.662 3.371251
#> 340  -0.661 3.372040
#> 341  -0.660 3.372830
#> 342  -0.659 3.373620
#> 343  -0.658 3.374410
#> 344  -0.657 3.375202
#> 345  -0.656 3.375994
#> 346  -0.655 3.376786
#> 347  -0.654 3.377579
#> 348  -0.653 3.378373
#> 349  -0.652 3.379167
#> 350  -0.651 3.379962
#> 351  -0.650 3.380758
#> 352  -0.649 3.381554
#> 353  -0.648 3.382351
#> 354  -0.647 3.383149
#> 355  -0.646 3.383947
#> 356  -0.645 3.384746
#> 357  -0.644 3.385545
#> 358  -0.643 3.386346
#> 359  -0.642 3.387146
#> 360  -0.641 3.387948
#> 361  -0.640 3.388750
#> 362  -0.639 3.389552
#> 363  -0.638 3.390356
#> 364  -0.637 3.391160
#> 365  -0.636 3.391964
#> 366  -0.635 3.392769
#> 367  -0.634 3.393575
#> 368  -0.633 3.394382
#> 369  -0.632 3.395189
#> 370  -0.631 3.395997
#> 371  -0.630 3.396805
#> 372  -0.629 3.397614
#> 373  -0.628 3.398424
#> 374  -0.627 3.399234
#> 375  -0.626 3.400045
#> 376  -0.625 3.400857
#> 377  -0.624 3.401669
#> 378  -0.623 3.402482
#> 379  -0.622 3.403295
#> 380  -0.621 3.404109
#> 381  -0.620 3.404924
#> 382  -0.619 3.405740
#> 383  -0.618 3.406556
#> 384  -0.617 3.407373
#> 385  -0.616 3.408190
#> 386  -0.615 3.409008
#> 387  -0.614 3.409827
#> 388  -0.613 3.410646
#> 389  -0.612 3.411466
#> 390  -0.611 3.412287
#> 391  -0.610 3.413108
#> 392  -0.609 3.413930
#> 393  -0.608 3.414753
#> 394  -0.607 3.415576
#> 395  -0.606 3.416400
#> 396  -0.605 3.417225
#> 397  -0.604 3.418050
#> 398  -0.603 3.418876
#> 399  -0.602 3.419702
#> 400  -0.601 3.420530
#> 401  -0.600 3.421358
#> 402  -0.599 3.422186
#> 403  -0.598 3.423015
#> 404  -0.597 3.423845
#> 405  -0.596 3.424676
#> 406  -0.595 3.425507
#> 407  -0.594 3.426339
#> 408  -0.593 3.427171
#> 409  -0.592 3.428004
#> 410  -0.591 3.428838
#> 411  -0.590 3.429673
#> 412  -0.589 3.430508
#> 413  -0.588 3.431344
#> 414  -0.587 3.432180
#> 415  -0.586 3.433017
#> 416  -0.585 3.433855
#> 417  -0.584 3.434693
#> 418  -0.583 3.435533
#> 419  -0.582 3.436372
#> 420  -0.581 3.437213
#> 421  -0.580 3.438054
#> 422  -0.579 3.438896
#> 423  -0.578 3.439738
#> 424  -0.577 3.440581
#> 425  -0.576 3.441425
#> 426  -0.575 3.442270
#> 427  -0.574 3.443115
#> 428  -0.573 3.443961
#> 429  -0.572 3.444807
#> 430  -0.571 3.445654
#> 431  -0.570 3.446502
#> 432  -0.569 3.447351
#> 433  -0.568 3.448200
#> 434  -0.567 3.449050
#> 435  -0.566 3.449900
#> 436  -0.565 3.450752
#> 437  -0.564 3.451603
#> 438  -0.563 3.452456
#> 439  -0.562 3.453309
#> 440  -0.561 3.454163
#> 441  -0.560 3.455018
#> 442  -0.559 3.455873
#> 443  -0.558 3.456729
#> 444  -0.557 3.457586
#> 445  -0.556 3.458443
#> 446  -0.555 3.459301
#> 447  -0.554 3.460160
#> 448  -0.553 3.461019
#> 449  -0.552 3.461879
#> 450  -0.551 3.462740
#> 451  -0.550 3.463601
#> 452  -0.549 3.464463
#> 453  -0.548 3.465326
#> 454  -0.547 3.466190
#> 455  -0.546 3.467054
#> 456  -0.545 3.467919
#> 457  -0.544 3.468784
#> 458  -0.543 3.469651
#> 459  -0.542 3.470518
#> 460  -0.541 3.471385
#> 461  -0.540 3.472253
#> 462  -0.539 3.473122
#> 463  -0.538 3.473992
#> 464  -0.537 3.474863
#> 465  -0.536 3.475734
#> 466  -0.535 3.476605
#> 467  -0.534 3.477478
#> 468  -0.533 3.478351
#> 469  -0.532 3.479225
#> 470  -0.531 3.480099
#> 471  -0.530 3.480975
#> 472  -0.529 3.481851
#> 473  -0.528 3.482727
#> 474  -0.527 3.483605
#> 475  -0.526 3.484483
#> 476  -0.525 3.485361
#> 477  -0.524 3.486241
#> 478  -0.523 3.487121
#> 479  -0.522 3.488002
#> 480  -0.521 3.488883
#> 481  -0.520 3.489766
#> 482  -0.519 3.490648
#> 483  -0.518 3.491532
#> 484  -0.517 3.492416
#> 485  -0.516 3.493302
#> 486  -0.515 3.494187
#> 487  -0.514 3.495074
#> 488  -0.513 3.495961
#> 489  -0.512 3.496849
#> 490  -0.511 3.497737
#> 491  -0.510 3.498627
#> 492  -0.509 3.499517
#> 493  -0.508 3.500407
#> 494  -0.507 3.501299
#> 495  -0.506 3.502191
#> 496  -0.505 3.503084
#> 497  -0.504 3.503977
#> 498  -0.503 3.504871
#> 499  -0.502 3.505766
#> 500  -0.501 3.506662
#> 501  -0.500 3.507559
#> 502  -0.499 3.508456
#> 503  -0.498 3.509353
#> 504  -0.497 3.510252
#> 505  -0.496 3.511151
#> 506  -0.495 3.512051
#> 507  -0.494 3.512952
#> 508  -0.493 3.513853
#> 509  -0.492 3.514755
#> 510  -0.491 3.515658
#> 511  -0.490 3.516562
#> 512  -0.489 3.517466
#> 513  -0.488 3.518371
#> 514  -0.487 3.519277
#> 515  -0.486 3.520183
#> 516  -0.485 3.521090
#> 517  -0.484 3.521998
#> 518  -0.483 3.522907
#> 519  -0.482 3.523816
#> 520  -0.481 3.524726
#> 521  -0.480 3.525637
#> 522  -0.479 3.526548
#> 523  -0.478 3.527461
#> 524  -0.477 3.528374
#> 525  -0.476 3.529287
#> 526  -0.475 3.530202
#> 527  -0.474 3.531117
#> 528  -0.473 3.532033
#> 529  -0.472 3.532949
#> 530  -0.471 3.533867
#> 531  -0.470 3.534785
#> 532  -0.469 3.535703
#> 533  -0.468 3.536623
#> 534  -0.467 3.537543
#> 535  -0.466 3.538464
#> 536  -0.465 3.539386
#> 537  -0.464 3.540308
#> 538  -0.463 3.541232
#> 539  -0.462 3.542155
#> 540  -0.461 3.543080
#> 541  -0.460 3.544006
#> 542  -0.459 3.544932
#> 543  -0.458 3.545859
#> 544  -0.457 3.546786
#> 545  -0.456 3.547714
#> 546  -0.455 3.548644
#> 547  -0.454 3.549573
#> 548  -0.453 3.550504
#> 549  -0.452 3.551435
#> 550  -0.451 3.552367
#> 551  -0.450 3.553300
#> 552  -0.449 3.554234
#> 553  -0.448 3.555168
#> 554  -0.447 3.556103
#> 555  -0.446 3.557039
#> 556  -0.445 3.557975
#> 557  -0.444 3.558912
#> 558  -0.443 3.559850
#> 559  -0.442 3.560789
#> 560  -0.441 3.561729
#> 561  -0.440 3.562669
#> 562  -0.439 3.563610
#> 563  -0.438 3.564552
#> 564  -0.437 3.565494
#> 565  -0.436 3.566437
#> 566  -0.435 3.567381
#> 567  -0.434 3.568326
#> 568  -0.433 3.569272
#> 569  -0.432 3.570218
#> 570  -0.431 3.571165
#> 571  -0.430 3.572113
#> 572  -0.429 3.573061
#> 573  -0.428 3.574010
#> 574  -0.427 3.574960
#> 575  -0.426 3.575911
#> 576  -0.425 3.576863
#> 577  -0.424 3.577815
#> 578  -0.423 3.578768
#> 579  -0.422 3.579722
#> 580  -0.421 3.580676
#> 581  -0.420 3.581632
#> 582  -0.419 3.582588
#> 583  -0.418 3.583545
#> 584  -0.417 3.584502
#> 585  -0.416 3.585461
#> 586  -0.415 3.586420
#> 587  -0.414 3.587380
#> 588  -0.413 3.588340
#> 589  -0.412 3.589302
#> 590  -0.411 3.590264
#> 591  -0.410 3.591227
#> 592  -0.409 3.592191
#> 593  -0.408 3.593155
#> 594  -0.407 3.594121
#> 595  -0.406 3.595087
#> 596  -0.405 3.596053
#> 597  -0.404 3.597021
#> 598  -0.403 3.597989
#> 599  -0.402 3.598959
#> 600  -0.401 3.599928
#> 601  -0.400 3.600899
#> 602  -0.399 3.601871
#> 603  -0.398 3.602843
#> 604  -0.397 3.603816
#> 605  -0.396 3.604789
#> 606  -0.395 3.605764
#> 607  -0.394 3.606739
#> 608  -0.393 3.607715
#> 609  -0.392 3.608692
#> 610  -0.391 3.609670
#> 611  -0.390 3.610648
#> 612  -0.389 3.611628
#> 613  -0.388 3.612608
#> 614  -0.387 3.613588
#> 615  -0.386 3.614570
#> 616  -0.385 3.615552
#> 617  -0.384 3.616535
#> 618  -0.383 3.617519
#> 619  -0.382 3.618504
#> 620  -0.381 3.619489
#> 621  -0.380 3.620475
#> 622  -0.379 3.621462
#> 623  -0.378 3.622450
#> 624  -0.377 3.623439
#> 625  -0.376 3.624428
#> 626  -0.375 3.625418
#> 627  -0.374 3.626409
#> 628  -0.373 3.627401
#> 629  -0.372 3.628394
#> 630  -0.371 3.629387
#> 631  -0.370 3.630381
#> 632  -0.369 3.631376
#> 633  -0.368 3.632372
#> 634  -0.367 3.633368
#> 635  -0.366 3.634366
#> 636  -0.365 3.635364
#> 637  -0.364 3.636362
#> 638  -0.363 3.637362
#> 639  -0.362 3.638363
#> 640  -0.361 3.639364
#> 641  -0.360 3.640366
#> 642  -0.359 3.641369
#> 643  -0.358 3.642372
#> 644  -0.357 3.643377
#> 645  -0.356 3.644382
#> 646  -0.355 3.645388
#> 647  -0.354 3.646395
#> 648  -0.353 3.647403
#> 649  -0.352 3.648411
#> 650  -0.351 3.649420
#> 651  -0.350 3.650430
#> 652  -0.349 3.651441
#> 653  -0.348 3.652453
#> 654  -0.347 3.653465
#> 655  -0.346 3.654479
#> 656  -0.345 3.655493
#> 657  -0.344 3.656508
#> 658  -0.343 3.657523
#> 659  -0.342 3.658540
#> 660  -0.341 3.659557
#> 661  -0.340 3.660575
#> 662  -0.339 3.661594
#> 663  -0.338 3.662614
#> 664  -0.337 3.663635
#> 665  -0.336 3.664656
#> 666  -0.335 3.665678
#> 667  -0.334 3.666701
#> 668  -0.333 3.667725
#> 669  -0.332 3.668749
#> 670  -0.331 3.669775
#> 671  -0.330 3.670801
#> 672  -0.329 3.671828
#> 673  -0.328 3.672856
#> 674  -0.327 3.673885
#> 675  -0.326 3.674914
#> 676  -0.325 3.675945
#> 677  -0.324 3.676976
#> 678  -0.323 3.678008
#> 679  -0.322 3.679041
#> 680  -0.321 3.680074
#> 681  -0.320 3.681109
#> 682  -0.319 3.682144
#> 683  -0.318 3.683180
#> 684  -0.317 3.684217
#> 685  -0.316 3.685255
#> 686  -0.315 3.686293
#> 687  -0.314 3.687332
#> 688  -0.313 3.688373
#> 689  -0.312 3.689414
#> 690  -0.311 3.690456
#> 691  -0.310 3.691498
#> 692  -0.309 3.692542
#> 693  -0.308 3.693586
#> 694  -0.307 3.694631
#> 695  -0.306 3.695677
#> 696  -0.305 3.696724
#> 697  -0.304 3.697772
#> 698  -0.303 3.698820
#> 699  -0.302 3.699870
#> 700  -0.301 3.700920
#> 701  -0.300 3.701971
#> 702  -0.299 3.703023
#> 703  -0.298 3.704075
#> 704  -0.297 3.705129
#> 705  -0.296 3.706183
#> 706  -0.295 3.707238
#> 707  -0.294 3.708294
#> 708  -0.293 3.709351
#> 709  -0.292 3.710409
#> 710  -0.291 3.711467
#> 711  -0.290 3.712527
#> 712  -0.289 3.713587
#> 713  -0.288 3.714648
#> 714  -0.287 3.715710
#> 715  -0.286 3.716773
#> 716  -0.285 3.717836
#> 717  -0.284 3.718901
#> 718  -0.283 3.719966
#> 719  -0.282 3.721032
#> 720  -0.281 3.722099
#> 721  -0.280 3.723167
#> 722  -0.279 3.724236
#> 723  -0.278 3.725305
#> 724  -0.277 3.726375
#> 725  -0.276 3.727447
#> 726  -0.275 3.728519
#> 727  -0.274 3.729592
#> 728  -0.273 3.730665
#> 729  -0.272 3.731740
#> 730  -0.271 3.732815
#> 731  -0.270 3.733892
#> 732  -0.269 3.734969
#> 733  -0.268 3.736047
#> 734  -0.267 3.737126
#> 735  -0.266 3.738206
#> 736  -0.265 3.739286
#> 737  -0.264 3.740368
#> 738  -0.263 3.741450
#> 739  -0.262 3.742533
#> 740  -0.261 3.743617
#> 741  -0.260 3.744702
#> 742  -0.259 3.745788
#> 743  -0.258 3.746874
#> 744  -0.257 3.747962
#> 745  -0.256 3.749050
#> 746  -0.255 3.750139
#> 747  -0.254 3.751230
#> 748  -0.253 3.752320
#> 749  -0.252 3.753412
#> 750  -0.251 3.754505
#> 751  -0.250 3.755598
#> 752  -0.249 3.756693
#> 753  -0.248 3.757788
#> 754  -0.247 3.758884
#> 755  -0.246 3.759981
#> 756  -0.245 3.761079
#> 757  -0.244 3.762178
#> 758  -0.243 3.763277
#> 759  -0.242 3.764378
#> 760  -0.241 3.765479
#> 761  -0.240 3.766581
#> 762  -0.239 3.767684
#> 763  -0.238 3.768788
#> 764  -0.237 3.769893
#> 765  -0.236 3.770999
#> 766  -0.235 3.772106
#> 767  -0.234 3.773213
#> 768  -0.233 3.774321
#> 769  -0.232 3.775431
#> 770  -0.231 3.776541
#> 771  -0.230 3.777652
#> 772  -0.229 3.778764
#> 773  -0.228 3.779876
#> 774  -0.227 3.780990
#> 775  -0.226 3.782104
#> 776  -0.225 3.783220
#> 777  -0.224 3.784336
#> 778  -0.223 3.785453
#> 779  -0.222 3.786571
#> 780  -0.221 3.787690
#> 781  -0.220 3.788810
#> 782  -0.219 3.789931
#> 783  -0.218 3.791052
#> 784  -0.217 3.792175
#> 785  -0.216 3.793298
#> 786  -0.215 3.794422
#> 787  -0.214 3.795547
#> 788  -0.213 3.796673
#> 789  -0.212 3.797800
#> 790  -0.211 3.798928
#> 791  -0.210 3.800057
#> 792  -0.209 3.801186
#> 793  -0.208 3.802317
#> 794  -0.207 3.803448
#> 795  -0.206 3.804581
#> 796  -0.205 3.805714
#> 797  -0.204 3.806848
#> 798  -0.203 3.807983
#> 799  -0.202 3.809119
#> 800  -0.201 3.810255
#> 801  -0.200 3.811393
#> 802  -0.199 3.812531
#> 803  -0.198 3.813671
#> 804  -0.197 3.814811
#> 805  -0.196 3.815953
#> 806  -0.195 3.817095
#> 807  -0.194 3.818238
#> 808  -0.193 3.819382
#> 809  -0.192 3.820527
#> 810  -0.191 3.821672
#> 811  -0.190 3.822819
#> 812  -0.189 3.823967
#> 813  -0.188 3.825115
#> 814  -0.187 3.826264
#> 815  -0.186 3.827415
#> 816  -0.185 3.828566
#> 817  -0.184 3.829718
#> 818  -0.183 3.830871
#> 819  -0.182 3.832025
#> 820  -0.181 3.833180
#> 821  -0.180 3.834336
#> 822  -0.179 3.835492
#> 823  -0.178 3.836650
#> 824  -0.177 3.837808
#> 825  -0.176 3.838968
#> 826  -0.175 3.840128
#> 827  -0.174 3.841289
#> 828  -0.173 3.842451
#> 829  -0.172 3.843614
#> 830  -0.171 3.844778
#> 831  -0.170 3.845943
#> 832  -0.169 3.847109
#> 833  -0.168 3.848276
#> 834  -0.167 3.849444
#> 835  -0.166 3.850612
#> 836  -0.165 3.851782
#> 837  -0.164 3.852952
#> 838  -0.163 3.854123
#> 839  -0.162 3.855296
#> 840  -0.161 3.856469
#> 841  -0.160 3.857643
#> 842  -0.159 3.858818
#> 843  -0.158 3.859994
#> 844  -0.157 3.861171
#> 845  -0.156 3.862349
#> 846  -0.155 3.863528
#> 847  -0.154 3.864707
#> 848  -0.153 3.865888
#> 849  -0.152 3.867069
#> 850  -0.151 3.868252
#> 851  -0.150 3.869435
#> 852  -0.149 3.870620
#> 853  -0.148 3.871805
#> 854  -0.147 3.872991
#> 855  -0.146 3.874178
#> 856  -0.145 3.875366
#> 857  -0.144 3.876555
#> 858  -0.143 3.877745
#> 859  -0.142 3.878936
#> 860  -0.141 3.880128
#> 861  -0.140 3.881321
#> 862  -0.139 3.882514
#> 863  -0.138 3.883709
#> 864  -0.137 3.884905
#> 865  -0.136 3.886101
#> 866  -0.135 3.887299
#> 867  -0.134 3.888497
#> 868  -0.133 3.889696
#> 869  -0.132 3.890897
#> 870  -0.131 3.892098
#> 871  -0.130 3.893300
#> 872  -0.129 3.894503
#> 873  -0.128 3.895707
#> 874  -0.127 3.896912
#> 875  -0.126 3.898118
#> 876  -0.125 3.899325
#> 877  -0.124 3.900533
#> 878  -0.123 3.901742
#> 879  -0.122 3.902951
#> 880  -0.121 3.904162
#> 881  -0.120 3.905374
#> 882  -0.119 3.906586
#> 883  -0.118 3.907800
#> 884  -0.117 3.909014
#> 885  -0.116 3.910230
#> 886  -0.115 3.911446
#> 887  -0.114 3.912664
#> 888  -0.113 3.913882
#> 889  -0.112 3.915101
#> 890  -0.111 3.916321
#> 891  -0.110 3.917543
#> 892  -0.109 3.918765
#> 893  -0.108 3.919988
#> 894  -0.107 3.921212
#> 895  -0.106 3.922437
#> 896  -0.105 3.923663
#> 897  -0.104 3.924890
#> 898  -0.103 3.926118
#> 899  -0.102 3.927347
#> 900  -0.101 3.928577
#> 901  -0.100 3.929807
#> 902  -0.099 3.931039
#> 903  -0.098 3.932272
#> 904  -0.097 3.933506
#> 905  -0.096 3.934740
#> 906  -0.095 3.935976
#> 907  -0.094 3.937212
#> 908  -0.093 3.938450
#> 909  -0.092 3.939689
#> 910  -0.091 3.940928
#> 911  -0.090 3.942169
#> 912  -0.089 3.943410
#> 913  -0.088 3.944652
#> 914  -0.087 3.945896
#> 915  -0.086 3.947140
#> 916  -0.085 3.948386
#> 917  -0.084 3.949632
#> 918  -0.083 3.950879
#> 919  -0.082 3.952127
#> 920  -0.081 3.953377
#> 921  -0.080 3.954627
#> 922  -0.079 3.955878
#> 923  -0.078 3.957130
#> 924  -0.077 3.958383
#> 925  -0.076 3.959638
#> 926  -0.075 3.960893
#> 927  -0.074 3.962149
#> 928  -0.073 3.963406
#> 929  -0.072 3.964664
#> 930  -0.071 3.965923
#> 931  -0.070 3.967183
#> 932  -0.069 3.968444
#> 933  -0.068 3.969706
#> 934  -0.067 3.970969
#> 935  -0.066 3.972233
#> 936  -0.065 3.973498
#> 937  -0.064 3.974764
#> 938  -0.063 3.976031
#> 939  -0.062 3.977299
#> 940  -0.061 3.978568
#> 941  -0.060 3.979838
#> 942  -0.059 3.981109
#> 943  -0.058 3.982381
#> 944  -0.057 3.983653
#> 945  -0.056 3.984927
#> 946  -0.055 3.986202
#> 947  -0.054 3.987478
#> 948  -0.053 3.988755
#> 949  -0.052 3.990033
#> 950  -0.051 3.991312
#> 951  -0.050 3.992592
#> 952  -0.049 3.993872
#> 953  -0.048 3.995154
#> 954  -0.047 3.996437
#> 955  -0.046 3.997721
#> 956  -0.045 3.999006
#> 957  -0.044 4.000292
#> 958  -0.043 4.001579
#> 959  -0.042 4.002866
#> 960  -0.041 4.004155
#> 961  -0.040 4.005445
#> 962  -0.039 4.006736
#> 963  -0.038 4.008028
#> 964  -0.037 4.009321
#> 965  -0.036 4.010615
#> 966  -0.035 4.011910
#> 967  -0.034 4.013206
#> 968  -0.033 4.014502
#> 969  -0.032 4.015800
#> 970  -0.031 4.017099
#> 971  -0.030 4.018399
#> 972  -0.029 4.019700
#> 973  -0.028 4.021002
#> 974  -0.027 4.022305
#> 975  -0.026 4.023609
#> 976  -0.025 4.024914
#> 977  -0.024 4.026220
#> 978  -0.023 4.027527
#> 979  -0.022 4.028835
#> 980  -0.021 4.030144
#> 981  -0.020 4.031455
#> 982  -0.019 4.032766
#> 983  -0.018 4.034078
#> 984  -0.017 4.035391
#> 985  -0.016 4.036705
#> 986  -0.015 4.038020
#> 987  -0.014 4.039337
#> 988  -0.013 4.040654
#> 989  -0.012 4.041972
#> 990  -0.011 4.043291
#> 991  -0.010 4.044612
#> 992  -0.009 4.045933
#> 993  -0.008 4.047255
#> 994  -0.007 4.048579
#> 995  -0.006 4.049903
#> 996  -0.005 4.051228
#> 997  -0.004 4.052555
#> 998  -0.003 4.053882
#> 999  -0.002 4.055211
#> 1000 -0.001 4.056541
#> 1001  0.000 4.057871
#> 1002  0.001 4.059203
#> 1003  0.002 4.060535
#> 1004  0.003 4.061869
#> 1005  0.004 4.063204
#> 1006  0.005 4.064539
#> 1007  0.006 4.065876
#> 1008  0.007 4.067214
#> 1009  0.008 4.068553
#> 1010  0.009 4.069893
#> 1011  0.010 4.071234
#> 1012  0.011 4.072576
#> 1013  0.012 4.073919
#> 1014  0.013 4.075263
#> 1015  0.014 4.076608
#> 1016  0.015 4.077954
#> 1017  0.016 4.079301
#> 1018  0.017 4.080649
#> 1019  0.018 4.081999
#> 1020  0.019 4.083349
#> 1021  0.020 4.084700
#> 1022  0.021 4.086053
#> 1023  0.022 4.087406
#> 1024  0.023 4.088760
#> 1025  0.024 4.090116
#> 1026  0.025 4.091473
#> 1027  0.026 4.092830
#> 1028  0.027 4.094189
#> 1029  0.028 4.095548
#> 1030  0.029 4.096909
#> 1031  0.030 4.098271
#> 1032  0.031 4.099634
#> 1033  0.032 4.100998
#> 1034  0.033 4.102363
#> 1035  0.034 4.103729
#> 1036  0.035 4.105096
#> 1037  0.036 4.106464
#> 1038  0.037 4.107833
#> 1039  0.038 4.109203
#> 1040  0.039 4.110575
#> 1041  0.040 4.111947
#> 1042  0.041 4.113320
#> 1043  0.042 4.114695
#> 1044  0.043 4.116070
#> 1045  0.044 4.117447
#> 1046  0.045 4.118825
#> 1047  0.046 4.120203
#> 1048  0.047 4.121583
#> 1049  0.048 4.122964
#> 1050  0.049 4.124346
#> 1051  0.050 4.125729
#> 1052  0.051 4.127113
#> 1053  0.052 4.128498
#> 1054  0.053 4.129884
#> 1055  0.054 4.131271
#> 1056  0.055 4.132660
#> 1057  0.056 4.134049
#> 1058  0.057 4.135439
#> 1059  0.058 4.136831
#> 1060  0.059 4.138223
#> 1061  0.060 4.139617
#> 1062  0.061 4.141012
#> 1063  0.062 4.142408
#> 1064  0.063 4.143804
#> 1065  0.064 4.145202
#> 1066  0.065 4.146601
#> 1067  0.066 4.148001
#> 1068  0.067 4.149402
#> 1069  0.068 4.150805
#> 1070  0.069 4.152208
#> 1071  0.070 4.153612
#> 1072  0.071 4.155018
#> 1073  0.072 4.156424
#> 1074  0.073 4.157832
#> 1075  0.074 4.159241
#> 1076  0.075 4.160650
#> 1077  0.076 4.162061
#> 1078  0.077 4.163473
#> 1079  0.078 4.164886
#> 1080  0.079 4.166300
#> 1081  0.080 4.167715
#> 1082  0.081 4.169132
#> 1083  0.082 4.170549
#> 1084  0.083 4.171967
#> 1085  0.084 4.173387
#> 1086  0.085 4.174808
#> 1087  0.086 4.176229
#> 1088  0.087 4.177652
#> 1089  0.088 4.179076
#> 1090  0.089 4.180501
#> 1091  0.090 4.181927
#> 1092  0.091 4.183354
#> 1093  0.092 4.184782
#> 1094  0.093 4.186212
#> 1095  0.094 4.187642
#> 1096  0.095 4.189074
#> 1097  0.096 4.190506
#> 1098  0.097 4.191940
#> 1099  0.098 4.193375
#> 1100  0.099 4.194811
#> 1101  0.100 4.196248
#> 1102  0.101 4.197686
#> 1103  0.102 4.199125
#> 1104  0.103 4.200565
#> 1105  0.104 4.202006
#> 1106  0.105 4.203449
#> 1107  0.106 4.204892
#> 1108  0.107 4.206337
#> 1109  0.108 4.207783
#> 1110  0.109 4.209230
#> 1111  0.110 4.210678
#> 1112  0.111 4.212127
#> 1113  0.112 4.213577
#> 1114  0.113 4.215028
#> 1115  0.114 4.216481
#> 1116  0.115 4.217934
#> 1117  0.116 4.219389
#> 1118  0.117 4.220845
#> 1119  0.118 4.222301
#> 1120  0.119 4.223759
#> 1121  0.120 4.225218
#> 1122  0.121 4.226679
#> 1123  0.122 4.228140
#> 1124  0.123 4.229602
#> 1125  0.124 4.231066
#> 1126  0.125 4.232530
#> 1127  0.126 4.233996
#> 1128  0.127 4.235463
#> 1129  0.128 4.236931
#> 1130  0.129 4.238400
#> 1131  0.130 4.239870
#> 1132  0.131 4.241341
#> 1133  0.132 4.242814
#> 1134  0.133 4.244287
#> 1135  0.134 4.245762
#> 1136  0.135 4.247238
#> 1137  0.136 4.248715
#> 1138  0.137 4.250193
#> 1139  0.138 4.251672
#> 1140  0.139 4.253152
#> 1141  0.140 4.254633
#> 1142  0.141 4.256116
#> 1143  0.142 4.257600
#> 1144  0.143 4.259084
#> 1145  0.144 4.260570
#> 1146  0.145 4.262057
#> 1147  0.146 4.263545
#> 1148  0.147 4.265035
#> 1149  0.148 4.266525
#> 1150  0.149 4.268016
#> 1151  0.150 4.269509
#> 1152  0.151 4.271003
#> 1153  0.152 4.272498
#> 1154  0.153 4.273994
#> 1155  0.154 4.275491
#> 1156  0.155 4.276989
#> 1157  0.156 4.278489
#> 1158  0.157 4.279989
#> 1159  0.158 4.281491
#> 1160  0.159 4.282994
#> 1161  0.160 4.284498
#> 1162  0.161 4.286003
#> 1163  0.162 4.287509
#> 1164  0.163 4.289016
#> 1165  0.164 4.290525
#> 1166  0.165 4.292035
#> 1167  0.166 4.293545
#> 1168  0.167 4.295057
#> 1169  0.168 4.296570
#> 1170  0.169 4.298085
#> 1171  0.170 4.299600
#> 1172  0.171 4.301116
#> 1173  0.172 4.302634
#> 1174  0.173 4.304153
#> 1175  0.174 4.305673
#> 1176  0.175 4.307194
#> 1177  0.176 4.308716
#> 1178  0.177 4.310239
#> 1179  0.178 4.311764
#> 1180  0.179 4.313290
#> 1181  0.180 4.314816
#> 1182  0.181 4.316344
#> 1183  0.182 4.317873
#> 1184  0.183 4.319404
#> 1185  0.184 4.320935
#> 1186  0.185 4.322468
#> 1187  0.186 4.324001
#> 1188  0.187 4.325536
#> 1189  0.188 4.327072
#> 1190  0.189 4.328609
#> 1191  0.190 4.330148
#> 1192  0.191 4.331687
#> 1193  0.192 4.333228
#> 1194  0.193 4.334770
#> 1195  0.194 4.336312
#> 1196  0.195 4.337857
#> 1197  0.196 4.339402
#> 1198  0.197 4.340948
#> 1199  0.198 4.342496
#> 1200  0.199 4.344045
#> 1201  0.200 4.345594
#> 1202  0.201 4.347145
#> 1203  0.202 4.348698
#> 1204  0.203 4.350251
#> 1205  0.204 4.351806
#> 1206  0.205 4.353361
#> 1207  0.206 4.354918
#> 1208  0.207 4.356476
#> 1209  0.208 4.358035
#> 1210  0.209 4.359596
#> 1211  0.210 4.361157
#> 1212  0.211 4.362720
#> 1213  0.212 4.364284
#> 1214  0.213 4.365849
#> 1215  0.214 4.367415
#> 1216  0.215 4.368982
#> 1217  0.216 4.370551
#> 1218  0.217 4.372121
#> 1219  0.218 4.373692
#> 1220  0.219 4.375264
#> 1221  0.220 4.376837
#> 1222  0.221 4.378411
#> 1223  0.222 4.379987
#> 1224  0.223 4.381564
#> 1225  0.224 4.383142
#> 1226  0.225 4.384721
#> 1227  0.226 4.386301
#> 1228  0.227 4.387882
#> 1229  0.228 4.389465
#> 1230  0.229 4.391049
#> 1231  0.230 4.392634
#> 1232  0.231 4.394220
#> 1233  0.232 4.395808
#> 1234  0.233 4.397396
#> 1235  0.234 4.398986
#> 1236  0.235 4.400577
#> 1237  0.236 4.402169
#> 1238  0.237 4.403762
#> 1239  0.238 4.405356
#> 1240  0.239 4.406952
#> 1241  0.240 4.408549
#> 1242  0.241 4.410147
#> 1243  0.242 4.411746
#> 1244  0.243 4.413347
#> 1245  0.244 4.414948
#> 1246  0.245 4.416551
#> 1247  0.246 4.418155
#> 1248  0.247 4.419760
#> 1249  0.248 4.421366
#> 1250  0.249 4.422974
#> 1251  0.250 4.424583
#> 1252  0.251 4.426192
#> 1253  0.252 4.427804
#> 1254  0.253 4.429416
#> 1255  0.254 4.431029
#> 1256  0.255 4.432644
#> 1257  0.256 4.434260
#> 1258  0.257 4.435877
#> 1259  0.258 4.437495
#> 1260  0.259 4.439115
#> 1261  0.260 4.440735
#> 1262  0.261 4.442357
#> 1263  0.262 4.443980
#> 1264  0.263 4.445604
#> 1265  0.264 4.447230
#> 1266  0.265 4.448857
#> 1267  0.266 4.450484
#> 1268  0.267 4.452114
#> 1269  0.268 4.453744
#> 1270  0.269 4.455375
#> 1271  0.270 4.457008
#> 1272  0.271 4.458642
#> 1273  0.272 4.460277
#> 1274  0.273 4.461913
#> 1275  0.274 4.463551
#> 1276  0.275 4.465189
#> 1277  0.276 4.466829
#> 1278  0.277 4.468470
#> 1279  0.278 4.470113
#> 1280  0.279 4.471756
#> 1281  0.280 4.473401
#> 1282  0.281 4.475047
#> 1283  0.282 4.476694
#> 1284  0.283 4.478342
#> 1285  0.284 4.479992
#> 1286  0.285 4.481643
#> 1287  0.286 4.483295
#> 1288  0.287 4.484948
#> 1289  0.288 4.486602
#> 1290  0.289 4.488258
#> 1291  0.290 4.489915
#> 1292  0.291 4.491573
#> 1293  0.292 4.493232
#> 1294  0.293 4.494893
#> 1295  0.294 4.496554
#> 1296  0.295 4.498217
#> 1297  0.296 4.499882
#> 1298  0.297 4.501547
#> 1299  0.298 4.503214
#> 1300  0.299 4.504881
#> 1301  0.300 4.506550
#> 1302  0.301 4.508221
#> 1303  0.302 4.509892
#> 1304  0.303 4.511565
#> 1305  0.304 4.513239
#> 1306  0.305 4.514914
#> 1307  0.306 4.516590
#> 1308  0.307 4.518268
#> 1309  0.308 4.519947
#> 1310  0.309 4.521627
#> 1311  0.310 4.523308
#> 1312  0.311 4.524991
#> 1313  0.312 4.526674
#> 1314  0.313 4.528359
#> 1315  0.314 4.530046
#> 1316  0.315 4.531733
#> 1317  0.316 4.533422
#> 1318  0.317 4.535112
#> 1319  0.318 4.536803
#> 1320  0.319 4.538495
#> 1321  0.320 4.540189
#> 1322  0.321 4.541883
#> 1323  0.322 4.543580
#> 1324  0.323 4.545277
#> 1325  0.324 4.546975
#> 1326  0.325 4.548675
#> 1327  0.326 4.550376
#> 1328  0.327 4.552078
#> 1329  0.328 4.553782
#> 1330  0.329 4.555487
#> 1331  0.330 4.557192
#> 1332  0.331 4.558900
#> 1333  0.332 4.560608
#> 1334  0.333 4.562318
#> 1335  0.334 4.564029
#> 1336  0.335 4.565741
#> 1337  0.336 4.567454
#> 1338  0.337 4.569169
#> 1339  0.338 4.570885
#> 1340  0.339 4.572602
#> 1341  0.340 4.574320
#> 1342  0.341 4.576040
#> 1343  0.342 4.577761
#> 1344  0.343 4.579483
#> 1345  0.344 4.581206
#> 1346  0.345 4.582931
#> 1347  0.346 4.584656
#> 1348  0.347 4.586384
#> 1349  0.348 4.588112
#> 1350  0.349 4.589841
#> 1351  0.350 4.591572
#> 1352  0.351 4.593304
#> 1353  0.352 4.595038
#> 1354  0.353 4.596772
#> 1355  0.354 4.598508
#> 1356  0.355 4.600245
#> 1357  0.356 4.601984
#> 1358  0.357 4.603723
#> 1359  0.358 4.605464
#> 1360  0.359 4.607206
#> 1361  0.360 4.608949
#> 1362  0.361 4.610694
#> 1363  0.362 4.612440
#> 1364  0.363 4.614187
#> 1365  0.364 4.615935
#> 1366  0.365 4.617685
#> 1367  0.366 4.619436
#> 1368  0.367 4.621188
#> 1369  0.368 4.622942
#> 1370  0.369 4.624696
#> 1371  0.370 4.626452
#> 1372  0.371 4.628209
#> 1373  0.372 4.629968
#> 1374  0.373 4.631728
#> 1375  0.374 4.633489
#> 1376  0.375 4.635251
#> 1377  0.376 4.637014
#> 1378  0.377 4.638779
#> 1379  0.378 4.640545
#> 1380  0.379 4.642312
#> 1381  0.380 4.644081
#> 1382  0.381 4.645851
#> 1383  0.382 4.647622
#> 1384  0.383 4.649394
#> 1385  0.384 4.651168
#> 1386  0.385 4.652943
#> 1387  0.386 4.654719
#> 1388  0.387 4.656497
#> 1389  0.388 4.658275
#> 1390  0.389 4.660055
#> 1391  0.390 4.661837
#> 1392  0.391 4.663619
#> 1393  0.392 4.665403
#> 1394  0.393 4.667188
#> 1395  0.394 4.668974
#> 1396  0.395 4.670762
#> 1397  0.396 4.672551
#> 1398  0.397 4.674341
#> 1399  0.398 4.676132
#> 1400  0.399 4.677925
#> 1401  0.400 4.679719
#> 1402  0.401 4.681514
#> 1403  0.402 4.683311
#> 1404  0.403 4.685109
#> 1405  0.404 4.686908
#> 1406  0.405 4.688708
#> 1407  0.406 4.690510
#> 1408  0.407 4.692313
#> 1409  0.408 4.694117
#> 1410  0.409 4.695923
#> 1411  0.410 4.697730
#> 1412  0.411 4.699538
#> 1413  0.412 4.701347
#> 1414  0.413 4.703158
#> 1415  0.414 4.704970
#> 1416  0.415 4.706783
#> 1417  0.416 4.708597
#> 1418  0.417 4.710413
#> 1419  0.418 4.712230
#> 1420  0.419 4.714048
#> 1421  0.420 4.715868
#> 1422  0.421 4.717689
#> 1423  0.422 4.719511
#> 1424  0.423 4.721335
#> 1425  0.424 4.723160
#> 1426  0.425 4.724986
#> 1427  0.426 4.726813
#> 1428  0.427 4.728642
#> 1429  0.428 4.730472
#> 1430  0.429 4.732303
#> 1431  0.430 4.734135
#> 1432  0.431 4.735969
#> 1433  0.432 4.737804
#> 1434  0.433 4.739641
#> 1435  0.434 4.741478
#> 1436  0.435 4.743317
#> 1437  0.436 4.745158
#> 1438  0.437 4.746999
#> 1439  0.438 4.748842
#> 1440  0.439 4.750686
#> 1441  0.440 4.752532
#> 1442  0.441 4.754379
#> 1443  0.442 4.756227
#> 1444  0.443 4.758076
#> 1445  0.444 4.759927
#> 1446  0.445 4.761779
#> 1447  0.446 4.763632
#> 1448  0.447 4.765487
#> 1449  0.448 4.767342
#> 1450  0.449 4.769200
#> 1451  0.450 4.771058
#> 1452  0.451 4.772918
#> 1453  0.452 4.774779
#> 1454  0.453 4.776641
#> 1455  0.454 4.778505
#> 1456  0.455 4.780370
#> 1457  0.456 4.782236
#> 1458  0.457 4.784104
#> 1459  0.458 4.785973
#> 1460  0.459 4.787843
#> 1461  0.460 4.789714
#> 1462  0.461 4.791587
#> 1463  0.462 4.793461
#> 1464  0.463 4.795337
#> 1465  0.464 4.797214
#> 1466  0.465 4.799092
#> 1467  0.466 4.800971
#> 1468  0.467 4.802852
#> 1469  0.468 4.804734
#> 1470  0.469 4.806617
#> 1471  0.470 4.808502
#> 1472  0.471 4.810388
#> 1473  0.472 4.812275
#> 1474  0.473 4.814163
#> 1475  0.474 4.816053
#> 1476  0.475 4.817944
#> 1477  0.476 4.819837
#> 1478  0.477 4.821731
#> 1479  0.478 4.823626
#> 1480  0.479 4.825522
#> 1481  0.480 4.827420
#> 1482  0.481 4.829319
#> 1483  0.482 4.831219
#> 1484  0.483 4.833121
#> 1485  0.484 4.835024
#> 1486  0.485 4.836928
#> 1487  0.486 4.838834
#> 1488  0.487 4.840741
#> 1489  0.488 4.842649
#> 1490  0.489 4.844559
#> 1491  0.490 4.846470
#> 1492  0.491 4.848382
#> 1493  0.492 4.850296
#> 1494  0.493 4.852210
#> 1495  0.494 4.854127
#> 1496  0.495 4.856044
#> 1497  0.496 4.857963
#> 1498  0.497 4.859883
#> 1499  0.498 4.861805
#> 1500  0.499 4.863727
#> 1501  0.500 4.865652
#> 1502  0.501 4.867577
#> 1503  0.502 4.869504
#> 1504  0.503 4.871432
#> 1505  0.504 4.873361
#> 1506  0.505 4.875292
#> 1507  0.506 4.877224
#> 1508  0.507 4.879158
#> 1509  0.508 4.881093
#> 1510  0.509 4.883029
#> 1511  0.510 4.884966
#> 1512  0.511 4.886905
#> 1513  0.512 4.888845
#> 1514  0.513 4.890786
#> 1515  0.514 4.892729
#> 1516  0.515 4.894673
#> 1517  0.516 4.896619
#> 1518  0.517 4.898565
#> 1519  0.518 4.900513
#> 1520  0.519 4.902463
#> 1521  0.520 4.904413
#> 1522  0.521 4.906366
#> 1523  0.522 4.908319
#> 1524  0.523 4.910274
#> 1525  0.524 4.912230
#> 1526  0.525 4.914187
#> 1527  0.526 4.916146
#> 1528  0.527 4.918106
#> 1529  0.528 4.920067
#> 1530  0.529 4.922030
#> 1531  0.530 4.923994
#> 1532  0.531 4.925960
#> 1533  0.532 4.927926
#> 1534  0.533 4.929894
#> 1535  0.534 4.931864
#> 1536  0.535 4.933835
#> 1537  0.536 4.935807
#> 1538  0.537 4.937780
#> 1539  0.538 4.939755
#> 1540  0.539 4.941731
#> 1541  0.540 4.943709
#> 1542  0.541 4.945688
#> 1543  0.542 4.947668
#> 1544  0.543 4.949649
#> 1545  0.544 4.951632
#> 1546  0.545 4.953616
#> 1547  0.546 4.955602
#> 1548  0.547 4.957589
#> 1549  0.548 4.959577
#> 1550  0.549 4.961567
#> 1551  0.550 4.963557
#> 1552  0.551 4.965550
#> 1553  0.552 4.967543
#> 1554  0.553 4.969538
#> 1555  0.554 4.971535
#> 1556  0.555 4.973532
#> 1557  0.556 4.975531
#> 1558  0.557 4.977532
#> 1559  0.558 4.979533
#> 1560  0.559 4.981536
#> 1561  0.560 4.983541
#> 1562  0.561 4.985546
#> 1563  0.562 4.987554
#> 1564  0.563 4.989562
#> 1565  0.564 4.991572
#> 1566  0.565 4.993583
#> 1567  0.566 4.995595
#> 1568  0.567 4.997609
#> 1569  0.568 4.999625
#> 1570  0.569 5.001641
#> 1571  0.570 5.003659
#> 1572  0.571 5.005678
#> 1573  0.572 5.007699
#> 1574  0.573 5.009721
#> 1575  0.574 5.011744
#> 1576  0.575 5.013769
#> 1577  0.576 5.015795
#> 1578  0.577 5.017822
#> 1579  0.578 5.019851
#> 1580  0.579 5.021881
#> 1581  0.580 5.023913
#> 1582  0.581 5.025945
#> 1583  0.582 5.027980
#> 1584  0.583 5.030015
#> 1585  0.584 5.032052
#> 1586  0.585 5.034090
#> 1587  0.586 5.036130
#> 1588  0.587 5.038171
#> 1589  0.588 5.040213
#> 1590  0.589 5.042257
#> 1591  0.590 5.044302
#> 1592  0.591 5.046348
#> 1593  0.592 5.048396
#> 1594  0.593 5.050445
#> 1595  0.594 5.052496
#> 1596  0.595 5.054548
#> 1597  0.596 5.056601
#> 1598  0.597 5.058655
#> 1599  0.598 5.060711
#> 1600  0.599 5.062769
#> 1601  0.600 5.064827
#> 1602  0.601 5.066887
#> 1603  0.602 5.068949
#> 1604  0.603 5.071012
#> 1605  0.604 5.073076
#> 1606  0.605 5.075141
#> 1607  0.606 5.077208
#> 1608  0.607 5.079276
#> 1609  0.608 5.081346
#> 1610  0.609 5.083417
#> 1611  0.610 5.085489
#> 1612  0.611 5.087563
#> 1613  0.612 5.089638
#> 1614  0.613 5.091714
#> 1615  0.614 5.093792
#> 1616  0.615 5.095871
#> 1617  0.616 5.097952
#> 1618  0.617 5.100034
#> 1619  0.618 5.102117
#> 1620  0.619 5.104202
#> 1621  0.620 5.106288
#> 1622  0.621 5.108375
#> 1623  0.622 5.110464
#> 1624  0.623 5.112554
#> 1625  0.624 5.114645
#> 1626  0.625 5.116738
#> 1627  0.626 5.118833
#> 1628  0.627 5.120928
#> 1629  0.628 5.123025
#> 1630  0.629 5.125124
#> 1631  0.630 5.127223
#> 1632  0.631 5.129324
#> 1633  0.632 5.131427
#> 1634  0.633 5.133531
#> 1635  0.634 5.135636
#> 1636  0.635 5.137743
#> 1637  0.636 5.139851
#> 1638  0.637 5.141960
#> 1639  0.638 5.144071
#> 1640  0.639 5.146183
#> 1641  0.640 5.148296
#> 1642  0.641 5.150411
#> 1643  0.642 5.152527
#> 1644  0.643 5.154645
#> 1645  0.644 5.156764
#> 1646  0.645 5.158884
#> 1647  0.646 5.161006
#> 1648  0.647 5.163129
#> 1649  0.648 5.165254
#> 1650  0.649 5.167380
#> 1651  0.650 5.169507
#> 1652  0.651 5.171636
#> 1653  0.652 5.173766
#> 1654  0.653 5.175897
#> 1655  0.654 5.178030
#> 1656  0.655 5.180164
#> 1657  0.656 5.182300
#> 1658  0.657 5.184437
#> 1659  0.658 5.186575
#> 1660  0.659 5.188715
#> 1661  0.660 5.190856
#> 1662  0.661 5.192998
#> 1663  0.662 5.195142
#> 1664  0.663 5.197287
#> 1665  0.664 5.199434
#> 1666  0.665 5.201582
#> 1667  0.666 5.203731
#> 1668  0.667 5.205882
#> 1669  0.668 5.208034
#> 1670  0.669 5.210188
#> 1671  0.670 5.212343
#> 1672  0.671 5.214499
#> 1673  0.672 5.216657
#> 1674  0.673 5.218816
#> 1675  0.674 5.220976
#> 1676  0.675 5.223138
#> 1677  0.676 5.225301
#> 1678  0.677 5.227466
#> 1679  0.678 5.229632
#> 1680  0.679 5.231799
#> 1681  0.680 5.233968
#> 1682  0.681 5.236138
#> 1683  0.682 5.238310
#> 1684  0.683 5.240483
#> 1685  0.684 5.242657
#> 1686  0.685 5.244833
#> 1687  0.686 5.247010
#> 1688  0.687 5.249188
#> 1689  0.688 5.251368
#> 1690  0.689 5.253549
#> 1691  0.690 5.255732
#> 1692  0.691 5.257916
#> 1693  0.692 5.260102
#> 1694  0.693 5.262288
#> 1695  0.694 5.264477
#> 1696  0.695 5.266666
#> 1697  0.696 5.268857
#> 1698  0.697 5.271050
#> 1699  0.698 5.273243
#> 1700  0.699 5.275439
#> 1701  0.700 5.277635
#> 1702  0.701 5.279833
#> 1703  0.702 5.282032
#> 1704  0.703 5.284233
#> 1705  0.704 5.286435
#> 1706  0.705 5.288639
#> 1707  0.706 5.290844
#> 1708  0.707 5.293050
#> 1709  0.708 5.295258
#> 1710  0.709 5.297467
#> 1711  0.710 5.299678
#> 1712  0.711 5.301889
#> 1713  0.712 5.304103
#> 1714  0.713 5.306317
#> 1715  0.714 5.308533
#> 1716  0.715 5.310751
#> 1717  0.716 5.312970
#> 1718  0.717 5.315190
#> 1719  0.718 5.317412
#> 1720  0.719 5.319635
#> 1721  0.720 5.321859
#> 1722  0.721 5.324085
#> 1723  0.722 5.326312
#> 1724  0.723 5.328541
#> 1725  0.724 5.330771
#> 1726  0.725 5.333003
#> 1727  0.726 5.335235
#> 1728  0.727 5.337470
#> 1729  0.728 5.339705
#> 1730  0.729 5.341942
#> 1731  0.730 5.344181
#> 1732  0.731 5.346421
#> 1733  0.732 5.348662
#> 1734  0.733 5.350904
#> 1735  0.734 5.353148
#> 1736  0.735 5.355394
#> 1737  0.736 5.357641
#> 1738  0.737 5.359889
#> 1739  0.738 5.362139
#> 1740  0.739 5.364390
#> 1741  0.740 5.366642
#> 1742  0.741 5.368896
#> 1743  0.742 5.371151
#> 1744  0.743 5.373408
#> 1745  0.744 5.375666
#> 1746  0.745 5.377925
#> 1747  0.746 5.380186
#> 1748  0.747 5.382448
#> 1749  0.748 5.384712
#> 1750  0.749 5.386977
#> 1751  0.750 5.389243
#> 1752  0.751 5.391511
#> 1753  0.752 5.393780
#> 1754  0.753 5.396051
#> 1755  0.754 5.398323
#> 1756  0.755 5.400596
#> 1757  0.756 5.402871
#> 1758  0.757 5.405147
#> 1759  0.758 5.407425
#> 1760  0.759 5.409704
#> 1761  0.760 5.411985
#> 1762  0.761 5.414266
#> 1763  0.762 5.416550
#> 1764  0.763 5.418834
#> 1765  0.764 5.421120
#> 1766  0.765 5.423408
#> 1767  0.766 5.425697
#> 1768  0.767 5.427987
#> 1769  0.768 5.430279
#> 1770  0.769 5.432572
#> 1771  0.770 5.434866
#> 1772  0.771 5.437162
#> 1773  0.772 5.439460
#> 1774  0.773 5.441758
#> 1775  0.774 5.444058
#> 1776  0.775 5.446360
#> 1777  0.776 5.448663
#> 1778  0.777 5.450967
#> 1779  0.778 5.453273
#> 1780  0.779 5.455580
#> 1781  0.780 5.457888
#> 1782  0.781 5.460198
#> 1783  0.782 5.462510
#> 1784  0.783 5.464823
#> 1785  0.784 5.467137
#> 1786  0.785 5.469452
#> 1787  0.786 5.471769
#> 1788  0.787 5.474088
#> 1789  0.788 5.476407
#> 1790  0.789 5.478729
#> 1791  0.790 5.481051
#> 1792  0.791 5.483375
#> 1793  0.792 5.485701
#> 1794  0.793 5.488027
#> 1795  0.794 5.490356
#> 1796  0.795 5.492685
#> 1797  0.796 5.495016
#> 1798  0.797 5.497349
#> 1799  0.798 5.499683
#> 1800  0.799 5.502018
#> 1801  0.800 5.504355
#> 1802  0.801 5.506693
#> 1803  0.802 5.509032
#> 1804  0.803 5.511373
#> 1805  0.804 5.513715
#> 1806  0.805 5.516059
#> 1807  0.806 5.518404
#> 1808  0.807 5.520751
#> 1809  0.808 5.523099
#> 1810  0.809 5.525448
#> 1811  0.810 5.527799
#> 1812  0.811 5.530151
#> 1813  0.812 5.532504
#> 1814  0.813 5.534859
#> 1815  0.814 5.537216
#> 1816  0.815 5.539574
#> 1817  0.816 5.541933
#> 1818  0.817 5.544293
#> 1819  0.818 5.546655
#> 1820  0.819 5.549019
#> 1821  0.820 5.551384
#> 1822  0.821 5.553750
#> 1823  0.822 5.556118
#> 1824  0.823 5.558487
#> 1825  0.824 5.560857
#> 1826  0.825 5.563229
#> 1827  0.826 5.565602
#> 1828  0.827 5.567977
#> 1829  0.828 5.570353
#> 1830  0.829 5.572731
#> 1831  0.830 5.575110
#> 1832  0.831 5.577490
#> 1833  0.832 5.579872
#> 1834  0.833 5.582255
#> 1835  0.834 5.584639
#> 1836  0.835 5.587025
#> 1837  0.836 5.589413
#> 1838  0.837 5.591802
#> 1839  0.838 5.594192
#> 1840  0.839 5.596583
#> 1841  0.840 5.598976
#> 1842  0.841 5.601371
#> 1843  0.842 5.603767
#> 1844  0.843 5.606164
#> 1845  0.844 5.608563
#> 1846  0.845 5.610963
#> 1847  0.846 5.613364
#> 1848  0.847 5.615767
#> 1849  0.848 5.618171
#> 1850  0.849 5.620577
#> 1851  0.850 5.622984
#> 1852  0.851 5.625393
#> 1853  0.852 5.627803
#> 1854  0.853 5.630214
#> 1855  0.854 5.632627
#> 1856  0.855 5.635041
#> 1857  0.856 5.637457
#> 1858  0.857 5.639874
#> 1859  0.858 5.642292
#> 1860  0.859 5.644712
#> 1861  0.860 5.647133
#> 1862  0.861 5.649556
#> 1863  0.862 5.651980
#> 1864  0.863 5.654405
#> 1865  0.864 5.656832
#> 1866  0.865 5.659261
#> 1867  0.866 5.661690
#> 1868  0.867 5.664121
#> 1869  0.868 5.666554
#> 1870  0.869 5.668988
#> 1871  0.870 5.671423
#> 1872  0.871 5.673860
#> 1873  0.872 5.676298
#> 1874  0.873 5.678738
#> 1875  0.874 5.681179
#> 1876  0.875 5.683621
#> 1877  0.876 5.686065
#> 1878  0.877 5.688510
#> 1879  0.878 5.690957
#> 1880  0.879 5.693405
#> 1881  0.880 5.695854
#> 1882  0.881 5.698305
#> 1883  0.882 5.700757
#> 1884  0.883 5.703211
#> 1885  0.884 5.705666
#> 1886  0.885 5.708122
#> 1887  0.886 5.710580
#> 1888  0.887 5.713039
#> 1889  0.888 5.715500
#> 1890  0.889 5.717962
#> 1891  0.890 5.720426
#> 1892  0.891 5.722891
#> 1893  0.892 5.725357
#> 1894  0.893 5.727825
#> 1895  0.894 5.730294
#> 1896  0.895 5.732765
#> 1897  0.896 5.735237
#> 1898  0.897 5.737710
#> 1899  0.898 5.740185
#> 1900  0.899 5.742661
#> 1901  0.900 5.745139
#> 1902  0.901 5.747618
#> 1903  0.902 5.750098
#> 1904  0.903 5.752580
#> 1905  0.904 5.755063
#> 1906  0.905 5.757548
#> 1907  0.906 5.760034
#> 1908  0.907 5.762522
#> 1909  0.908 5.765010
#> 1910  0.909 5.767501
#> 1911  0.910 5.769992
#> 1912  0.911 5.772486
#> 1913  0.912 5.774980
#> 1914  0.913 5.777476
#> 1915  0.914 5.779973
#> 1916  0.915 5.782472
#> 1917  0.916 5.784972
#> 1918  0.917 5.787474
#> 1919  0.918 5.789977
#> 1920  0.919 5.792481
#> 1921  0.920 5.794987
#> 1922  0.921 5.797494
#> 1923  0.922 5.800003
#> 1924  0.923 5.802513
#> 1925  0.924 5.805024
#> 1926  0.925 5.807537
#> 1927  0.926 5.810051
#> 1928  0.927 5.812567
#> 1929  0.928 5.815084
#> 1930  0.929 5.817602
#> 1931  0.930 5.820122
#> 1932  0.931 5.822644
#> 1933  0.932 5.825166
#> 1934  0.933 5.827690
#> 1935  0.934 5.830216
#> 1936  0.935 5.832743
#> 1937  0.936 5.835271
#> 1938  0.937 5.837801
#> 1939  0.938 5.840332
#> 1940  0.939 5.842864
#> 1941  0.940 5.845398
#> 1942  0.941 5.847934
#> 1943  0.942 5.850470
#> 1944  0.943 5.853008
#> 1945  0.944 5.855548
#> 1946  0.945 5.858089
#> 1947  0.946 5.860631
#> 1948  0.947 5.863175
#> 1949  0.948 5.865720
#> 1950  0.949 5.868267
#> 1951  0.950 5.870815
#> 1952  0.951 5.873364
#> 1953  0.952 5.875915
#> 1954  0.953 5.878467
#> 1955  0.954 5.881021
#> 1956  0.955 5.883576
#> 1957  0.956 5.886132
#> 1958  0.957 5.888690
#> 1959  0.958 5.891249
#> 1960  0.959 5.893810
#> 1961  0.960 5.896372
#> 1962  0.961 5.898935
#> 1963  0.962 5.901500
#> 1964  0.963 5.904066
#> 1965  0.964 5.906634
#> 1966  0.965 5.909203
#> 1967  0.966 5.911773
#> 1968  0.967 5.914345
#> 1969  0.968 5.916918
#> 1970  0.969 5.919493
#> 1971  0.970 5.922069
#> 1972  0.971 5.924646
#> 1973  0.972 5.927225
#> 1974  0.973 5.929805
#> 1975  0.974 5.932387
#> 1976  0.975 5.934970
#> 1977  0.976 5.937554
#> 1978  0.977 5.940140
#> 1979  0.978 5.942727
#> 1980  0.979 5.945316
#> 1981  0.980 5.947906
#> 1982  0.981 5.950497
#> 1983  0.982 5.953090
#> 1984  0.983 5.955684
#> 1985  0.984 5.958280
#> 1986  0.985 5.960877
#> 1987  0.986 5.963475
#> 1988  0.987 5.966075
#> 1989  0.988 5.968677
#> 1990  0.989 5.971279
#> 1991  0.990 5.973883
#> 1992  0.991 5.976489
#> 1993  0.992 5.979095
#> 1994  0.993 5.981704
#> 1995  0.994 5.984313
#> 1996  0.995 5.986924
#> 1997  0.996 5.989536
#> 1998  0.997 5.992150
#> 1999  0.998 5.994765
#> 2000  0.999 5.997382
#> 2001  1.000 6.000000