Estimator accuracy and bias

A simulated dataset with rmse and mean absolute error for different approaches to estimating the true (asymptotic) Hill diversity based on finite samples from simulated species abundance distributions. Hill diversity was estimated with the naïve estimator, rarity(sample_freq, l), the Chao and Jost 2015 non-parametric asymptotic estimator Chao_Hill_abu(sample_freq, l), and God's estimator GUE(sample_freq, true_freq, l).

errs

Format

A data.frame with 2376 rows and 8 variables:

ell: Scalar, scaling parameter for Hill diversity

distribution Factor, family, one of "lnorm" or "gamma"

fitted.parameter Factor, shape parameter for fitting SAD

n: Numeric, individuals sampled

SAD: index of SAD sampled

estimator Character, estimator whose error is described

rmse Numeric, root mean squared error of estimator

bias Numeric, mean difference between estimator and true diversity

References

Chao A, Jost L (2015). “Estimating diversity and entropy profiles via discovery rates of new species.” Methods in Ecology and Evolution, 6(8), 873--882. ISSN 2041210X, doi: 10.1111/2041-21 .