Estimate ENS based on Hurlbert rarefaction

Implements estimate described in Dauby and Hardy 2011 for a class of rarefaction-based ENS diversity estimates. These estimates suffer from minimal bias and are quite efficient, while retaining some of the nice properties of Hill diversity metrics. They are parameterized by sample size k, and when k == 2 they are equivalent to Hill-Simpson diversity. One interpretation is that this ENS is the number of species in a perfectly even assemblage that would have the same rarefied richness as the focal assemblage/sample. Larger k values emphasize rare species, and as k approaches community size the Hulbert ENS approaches true richness. Unbiased estimators are given for k < sample size.

hurl(ab, k, maxit = 1e+05, tol = 1e-12)

Arguments

ab

A numeric vector of species abundances or relative abundances.

k

integer sample size parameter for rarefaction

maxit

integer, maximum number of iterations

tol

numeric, threshold for convergence

Value

Numeric scalar: estimated Hurlbert ENS

References

Dauby G, Hardy OJ (2012). “Sampled-based estimation of diversity sensu stricto by transforming Hurlbert diversities into effective number of species.” Ecography, 35(7), 661--672. doi: 10.1111/j.1600-0587.2011.06860.x . Hurlbert SH (1971). “The nonconcept of species diversity: a critique and alternative parameters.” Ecology, 52(4), 577--586. doi: 10.2307/1934145 , http://www.esajournals.org/doi/abs/10.2307/1934145.

Examples

ab = sample(10:50, 50, replace =TRUE)
hurl(ab, 2)
#> [1] 45.62047
# not run
# hurl(ab, 1e5) # returns an error