Sampling variability in Hill diversity

A simulated dataset with mean absolute error and log sd for sample and asymptotic Hill diversity estimates based on finite samples from simulated species abundance distributions. Hill diversity was estimated with the naïve estimator, rarity(sample_freq, l) and the Chao and Jost 2015 non-parametric asymptotic estimator Chao_Hill_abu(sample_freq, l), and parameterized by richness and evenness

plot_data

Format

A data.frame with 7224 rows and 9 variables:

evenness Factor, approximately the ratio of Hill-Simpson to richness

distribution Factor, family, one of "lnorm" or "gamma"

ell: Scalar, scaling parameter for Hill diversity

SS Integer, number of individuals sampled

sample: Numeric, Hill diversity of the sample

asymptotic: Numeric, estimated Hill diversity of the SAD based on asymptotic estimator

true_diversity Numeric, Hill diversity of the SAD

mean_sample Numeric, mean of sample diversity for combination of ell, evenness, distribution, and SS.

mean_asymptotic Numeric, mean of estimated asymptotic diversity for combination of ell, evenness, distribution, and SS.

References

Chao A, Jost L (2015). “Estimating diversity and entropy profiles via discovery rates of new species.” Methods in Ecology and Evolution, 6(8), 873--882. ISSN 2041210X, doi: 10.1111/2041-21 . Chao A, Ricotta C, Aldo P (2019). “Quantifying evenness and linking it to diversity, beta diversity, and similarity.” Ecology, 1--15. ISSN 0012-9658, doi: 10.1002/ecy.2852 .